+++ tracee-rules/process_tree/events_processing_test.go
@@ -1,0 +1,129 @@
+package process_tree
+
+import (
+	"github.com/aquasecurity/tracee/pkg/external"
+	"github.com/aquasecurity/tracee/tracee-rules/types"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+	"testing"
+)
+
+const TestContainerID = "a7f965fba4e145e02c99b1577febe0cb723a943d850278365994ac9b0190540e"
+
+func TestProcessTree_ProcessEvent(t *testing.T) {
+	tree := ProcessTree{
+		containers: map[string]*containerProcessTree{},
+		tree:       map[int]*types.ProcessInfo{},
+	}
+
+	ptid := 22482
+	ppid := 22447
+	forkTimestamp := 1639044471927303690
+	execCmd := []string{"ls"}
+	execBinaryPath := "/bin/busybox"
+	execBinaryCtime := uint(1625759227634052514)
+
+	forkEvent := external.Event{
+		Timestamp:           forkTimestamp,
+		ProcessID:           0,
+		ThreadID:            0,
+		ParentProcessID:     22422,
+		HostProcessID:       ppid,
+		HostThreadID:        ppid,
+		HostParentProcessID: 22422,
+		UserID:              0,
+		MountNS:             4026532548,
+		PIDNS:               4026532551,
+		ProcessName:         "sh",
+		HostName:            "aac1fa476fcd",
+		ContainerID:         TestContainerID,
+		EventID:             1002,
+		EventName:           "sched_process_fork",
+		ArgsNum:             4,
+		ReturnValue:         0,
+		StackAddresses:      nil,
+		Args: []external.Argument{
+			{ArgMeta: external.ArgMeta{Name: "parent_tid", Type: "int"}, Value: int32(ppid)},
+			{ArgMeta: external.ArgMeta{Name: "parent_ns_tid", Type: "int"}, Value: int32(0)},
+			{ArgMeta: external.ArgMeta{Name: "parent_pid", Type: "int"}, Value: int32(ppid)},
+			{ArgMeta: external.ArgMeta{Name: "parent_ns_pid", Type: "int"}, Value: int32(0)},
+			{ArgMeta: external.ArgMeta{Name: "child_tid", Type: "int"}, Value: int32(ptid)},
+			{ArgMeta: external.ArgMeta{Name: "child_ns_tid", Type: "int"}, Value: int32(0)},
+			{ArgMeta: external.ArgMeta{Name: "child_pid", Type: "int"}, Value: int32(ptid)},
+			{ArgMeta: external.ArgMeta{Name: "child_ns_pid", Type: "int"}, Value: int32(0)},
+		},
+	}
+	execEvent := external.Event{
+		Timestamp:           1639044471927556667,
+		ProcessID:           0,
+		ThreadID:            0,
+		ParentProcessID:     0,
+		HostProcessID:       ptid,
+		HostThreadID:        ptid,
+		HostParentProcessID: ppid,
+		UserID:              0,
+		MountNS:             4026532548,
+		PIDNS:               4026532551,
+		ProcessName:         "ls",
+		HostName:            "aac1fa476fcd",
+		ContainerID:         TestContainerID,
+		EventID:             1003,
+		EventName:           "sched_process_exec",
+		ArgsNum:             9,
+		ReturnValue:         0,
+		StackAddresses:      nil,
+		Args: []external.Argument{
+			{ArgMeta: external.ArgMeta{Name: "cmdpath", Type: "const char*"}, Value: interface{}("/bin/ls")},
+			{ArgMeta: external.ArgMeta{Name: "argv", Type: "const char**"}, Value: interface{}(execCmd)},
+			{ArgMeta: external.ArgMeta{Name: "env", Type: "const char**"}, Value: interface{}([]string{"HOSTNAME=aac1fa454fcd", "SHLVL=1", "HOME=/root", "TERM=xterm", "PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin", "PWD=/"})},
+			{ArgMeta: external.ArgMeta{Name: "pathname", Type: "const char*"}, Value: interface{}(execBinaryPath)},
+			{ArgMeta: external.ArgMeta{Name: "dev", Type: "dev_t"}, Value: interface{}(46)},
+			{ArgMeta: external.ArgMeta{Name: "inode", Type: "unsigned long"}, Value: interface{}(576807)},
+			{ArgMeta: external.ArgMeta{Name: "invoked_from_kernel", Type: "int"}, Value: interface{}(0)},
+			{ArgMeta: external.ArgMeta{Name: "ctime", Type: "unsigned long"}, Value: interface{}(uint64(execBinaryCtime))},
+			{ArgMeta: external.ArgMeta{Name: "sha256", Type: "const char*"}, Value: interface{}("abfd081fd7fad08d4743443061a12ebfbd25e3c5e446441795d472c389444527")},
+		},
+	}
+	exitEvent := external.Event{
+		Timestamp:           1639044471928009089,
+		ProcessID:           0,
+		ThreadID:            0,
+		ParentProcessID:     0,
+		HostProcessID:       ptid,
+		HostThreadID:        ptid,
+		HostParentProcessID: ppid,
+		UserID:              0,
+		MountNS:             4026532548,
+		PIDNS:               4026532551,
+		ProcessName:         "ls",
+		HostName:            "aac1fa476fcd",
+		ContainerID:         TestContainerID,
+		EventID:             1004,
+		EventName:           "sched_process_exit",
+		ArgsNum:             1,
+		ReturnValue:         0,
+		StackAddresses:      nil,
+		Args: []external.Argument{
+			{ArgMeta: external.ArgMeta{Name: "exit_code", Type: "long"}, Value: 0},
+		},
+	}
+
+	err := tree.ProcessEvent(forkEvent)
+	require.NoError(t, err)
+	err = tree.ProcessEvent(execEvent)
+	require.NoError(t, err)
+	process, err := tree.GetProcessInfo(ptid)
+	assert.NoError(t, err)
+	assert.Equal(t, ptid, process.InHostIDs.Tid)
+	assert.Equal(t, ptid, process.InHostIDs.Pid)
+	assert.Equal(t, ppid, process.InHostIDs.Ppid)
+	assert.Equal(t, forkTimestamp, process.StartTime)
+	assert.Equal(t, execCmd, process.Cmd)
+	assert.Equal(t, execBinaryPath, process.ExecutionBinary.Path)
+	assert.Equal(t, execBinaryCtime, process.ExecutionBinary.Ctime)
+	err = tree.processExit(exitEvent)
+	tree.EmptyProcessCache()
+	require.NoError(t, err)
+	_, err = tree.GetProcessInfo(ptid)
+	assert.Error(t, err)
+}
